Rename AddJobResourceDecorator2 to CreateJobResourceFromImageDecorator
The AddJobResourceDecorator and AddJobResourceDecorator2 classes are meant to fulfill different purposes, so it makes more sense to rename AddJobResourceDecorator2 to CreateJobResourceFromImageDecorator.

Fix https://github.com/quarkusio/quarkus/issues/31723
